refactor(web): extract ops/hooks/webhooks/pairing mixins from web_server.py (shard s4)
