+++ hermes_cli/web_server.py
@@ -11770,255 +11770,26 @@ def _mcp_install_action_name(name: str) -> str:
 _ACTION_LOG_FILES.setdefault("computer-use-grant", "action-computer-use-grant.log")
 
 
-# ---------------------------------------------------------------------------
-# Pairing endpoints — approve / revoke / list messaging pairing codes.
-#
-# These are how a remote admin onboards messaging users (Telegram, Discord, …)
-# without shell access.  Wraps gateway.pairing.PairingStore directly.
-# ---------------------------------------------------------------------------
+from hermes_cli.web_routers import pairing as _pairing_routes  # noqa: E402
 
+app.include_router(_pairing_routes.router)
+from hermes_cli.web_routers.pairing import (  # noqa: E402,F401 — legacy re-exports; tests call these via web_server.<name>
+    list_pairing,
+    approve_pairing,
+    revoke_pairing,
+    clear_pending_pairing,
+)
 
-def _pairing_store(profile: Optional[str] = None):
-    """Pairing store for ``profile`` — the dashboard's own when unspecified.
+from hermes_cli.web_routers import webhooks as _webhooks_routes  # noqa: E402
 
-    Every other admin endpoint scopes by profile, and the gateway already
-    keeps one store per served profile (``gateway/run.py``). Without this the
-    dashboard and desktop always read the global store, so an operator on a
-    named profile approves into a whitelist their gateway never consults.
-
-    ``PairingStore`` resolves the profile's home itself (``default`` maps back
-    to the global store), so this only needs to validate the name — no
-    ``_profile_scope`` needed, and nothing process-global is swapped across
-    the ``await`` boundary.
-    """
-    from gateway.pairing import PairingStore
-
-    requested = (profile or "").strip()
-    if not requested or requested.lower() == "current":
-        return PairingStore()
-
-    _resolve_profile_dir(requested)  # 400/404 on an unknown profile
-
-    return PairingStore(profile=requested)
-
-
-@app.get("/api/pairing")
-async def list_pairing(profile: Optional[str] = None):
-    store = _pairing_store(profile)
-    return {
-        "pending": store.list_pending(),
-        "approved": store.list_approved(),
-    }
-
-
-@app.post("/api/pairing/approve")
-async def approve_pairing(body: PairingApprove):
-    store = _pairing_store(body.profile)
-    platform = (body.platform or "").lower().strip()
-    # `request_id` is what an admin surface sends after listing pending
-    # requests; `code` is the one-time code the user relays from their DM.
-    # A GUI that only knows the older field name still works — a value with
-    # request-id shape routes to the request path either way.
-    target = (body.request_id or body.code or "").strip()
-    if not platform or not target:
-        raise HTTPException(
-            status_code=400, detail="platform and request_id or code are required"
-        )
-
-    by_request_id = bool(body.request_id) or store.looks_like_request_id(target)
-    if by_request_id:
-        result = store.approve_request(platform, target)
-    else:
-        result = store.approve_code(platform, target.upper())
-
-    if result:
-        return {"ok": True, "user": result}
-    # Lockout only gates the code path, so only report it there — otherwise a
-    # stale request id would surface as a bogus 429 while the platform sat
-    # locked out for an unrelated reason.
-    if not by_request_id and store._is_locked_out(platform):
-        raise HTTPException(
-            status_code=429,
-            detail=f"Platform '{platform}' is locked out after too many failed approvals.",
-        )
-    raise HTTPException(
-        status_code=404,
-        detail=f"Pairing request or code not found or expired for platform '{platform}'.",
-    )
-
-
-@app.post("/api/pairing/revoke")
-async def revoke_pairing(body: PairingRevoke):
-    store = _pairing_store(body.profile)
-    platform = (body.platform or "").lower().strip()
-    if not platform or not body.user_id:
-        raise HTTPException(status_code=400, detail="platform and user_id are required")
-    if store.revoke(platform, body.user_id):
-        return {"ok": True}
-    raise HTTPException(
-        status_code=404,
-        detail=f"User {body.user_id} not found in approved list for {platform}.",
-    )
-
-
-@app.post("/api/pairing/clear-pending")
-async def clear_pending_pairing(profile: Optional[str] = None):
-    store = _pairing_store(profile)
-    count = store.clear_pending()
-    return {"ok": True, "cleared": count}
-
-
-# ---------------------------------------------------------------------------
-# Webhook subscription endpoints — list / subscribe / remove.
-#
-# Wraps the same JSON store the CLI uses (hermes_cli.webhook); the webhook
-# adapter hot-reloads it without a gateway restart.  Per-route HMAC secrets
-# are redacted on read and surfaced once on create.
-# ---------------------------------------------------------------------------
-
-
-def _webhook_route_summary(name: str, route: Dict[str, Any], base_url: str) -> Dict[str, Any]:
-    return {
-        "name": name,
-        "description": route.get("description", ""),
-        "events": list(route.get("events") or []),
-        "deliver": route.get("deliver", "log"),
-        "deliver_only": bool(route.get("deliver_only")),
-        "prompt": route.get("prompt", ""),
-        "script": route.get("script", ""),
-        "skills": list(route.get("skills") or []),
-        "created_at": route.get("created_at"),
-        "url": f"{base_url}/webhooks/{name}",
-        # Secret is masked on read; full value only returned on create.
-        "secret_set": bool(route.get("secret")),
-        # Default-enabled; only an explicit enabled:false turns a route off.
-        "enabled": route.get("enabled", True) is not False,
-    }
-
-
-@app.get("/api/webhooks")
-async def list_webhooks():
-    import hermes_cli.webhook as wh
-
-    base_url = wh._get_webhook_base_url()
-    subs = wh._load_subscriptions()
-    return {
-        "enabled": wh._is_webhook_enabled(),
-        "base_url": base_url,
-        "subscriptions": [
-            _webhook_route_summary(name, route, base_url)
-            for name, route in subs.items()
-        ],
-    }
-
-
-@app.post("/api/webhooks/enable")
-async def enable_webhooks():
-    try:
-        _write_platform_enabled("webhook", True)
-    except Exception as exc:
-        _log.exception("Failed to enable webhook platform from dashboard")
-        raise HTTPException(
-            status_code=500,
-            detail="Failed to enable webhook platform.",
-        ) from exc
-
-    restart_result = _restart_gateway_after_webhook_enable()
-    return {
-        "ok": True,
-        "platform": "webhook",
-        "enabled": True,
-        "needs_restart": not restart_result["restart_started"],
-        **restart_result,
-    }
-
-
-@app.post("/api/webhooks")
-async def create_webhook(body: WebhookCreate):
-    import re as _re
-    import secrets as _secrets
-    import time as _time
-    import hermes_cli.webhook as wh
-
-    if not wh._is_webhook_enabled():
-        raise HTTPException(
-            status_code=400,
-            detail="Webhook platform is not enabled. Enable it from the Webhooks page first.",
-        )
-
-    name = (body.name or "").strip().lower().replace(" ", "-")
-    if not _re.match(r"^[a-z0-9][a-z0-9_-]*$", name):
-        raise HTTPException(
-            status_code=400,
-            detail="Invalid name. Use lowercase alphanumeric with hyphens/underscores.",
-        )
-
-    if body.deliver_only and body.deliver == "log":
-        raise HTTPException(
-            status_code=400,
-            detail="Direct delivery requires a real target (telegram, discord, …), not 'log'.",
-        )
-
-    secret = body.secret or _secrets.token_urlsafe(32)
-    route: Dict[str, Any] = {
-        "description": body.description or f"Dashboard-created subscription: {name}",
-        "events": [e.strip() for e in body.events if e.strip()],
-        "secret": secret,
-        "prompt": body.prompt or "",
-        "skills": [s.strip() for s in body.skills if s.strip()],
-        "deliver": body.deliver or "log",
-        "created_at": _time.strftime("%Y-%m-%dT%H:%M:%SZ", _time.gmtime()),
-    }
-    if body.script and body.script.strip():
-        route["script"] = body.script.strip()
-    if body.deliver_only:
-        route["deliver_only"] = True
-    if body.deliver_chat_id:
-        route["deliver_extra"] = {"chat_id": body.deliver_chat_id}
-
-    subs = wh._load_subscriptions()
-    subs[name] = route
-    wh._save_subscriptions(subs)
-
-    base_url = wh._get_webhook_base_url()
-    summary = _webhook_route_summary(name, route, base_url)
-    # Surface the secret exactly once, on create.
-    summary["secret"] = secret
-    return summary
-
-
-@app.delete("/api/webhooks/{name}")
-async def delete_webhook(name: str):
-    import hermes_cli.webhook as wh
-
-    key = (name or "").strip().lower()
-    subs = wh._load_subscriptions()
-    if key not in subs:
-        raise HTTPException(status_code=404, detail=f"No subscription named '{key}'")
-    del subs[key]
-    wh._save_subscriptions(subs)
-    return {"ok": True}
-
-
-@app.put("/api/webhooks/{name}/enabled")
-async def set_webhook_enabled(name: str, body: WebhookEnabledToggle):
-    """Enable or disable a webhook route.
-
-    Disabled routes stay in the subscriptions file (so they can be
-    re-enabled) but the gateway rejects incoming events with 403.  The
-    gateway hot-reloads the subscriptions file, so this takes effect on the
-    next event without a restart.
-    """
-    import hermes_cli.webhook as wh
-
-    key = (name or "").strip().lower()
-    subs = wh._load_subscriptions()
-    if key not in subs:
-        raise HTTPException(status_code=404, detail=f"No subscription named '{key}'")
-    subs[key]["enabled"] = bool(body.enabled)
-    wh._save_subscriptions(subs)
-    return {"ok": True, "name": key, "enabled": bool(body.enabled)}
-
+app.include_router(_webhooks_routes.router)
+from hermes_cli.web_routers.webhooks import (  # noqa: E402,F401 — legacy re-exports; tests call these via web_server.<name>
+    list_webhooks,
+    enable_webhooks,
+    create_webhook,
+    delete_webhook,
+    set_webhook_enabled,
+)
 
 # ---------------------------------------------------------------------------
 # Gateway lifecycle endpoints — start / stop.
@@ -12288,388 +12059,28 @@ async def reset_memory(body: MemoryReset):
                 raise HTTPException(status_code=500, detail=f"Could not delete {fname}: {exc}")
     return {"ok": True, "deleted": deleted}
 
+from hermes_cli.web_routers import ops as _ops_routes  # noqa: E402
 
-# ---------------------------------------------------------------------------
-# Operations endpoints — doctor / security audit / backup / import /
-# checkpoints / hooks.
-#
-# Diagnostic and maintenance commands.  The long-running / text-output ones
-# (doctor, security audit, backup, import, skills install) are spawned as
-# background actions whose logs the dashboard tails via
-# /api/actions/{name}/status — same pattern as gateway restart and update.
-# The cheap, structured reads (hooks list, checkpoints list) return JSON
-# directly.
-# ---------------------------------------------------------------------------
+app.include_router(_ops_routes.router)
+from hermes_cli.web_routers.ops import (  # noqa: E402,F401 — legacy re-exports; tests call these via web_server.<name>
+    run_doctor,
+    run_security_audit,
+    run_backup,
+    download_dashboard_backup,
+    run_import,
+    run_import_upload,
+    list_checkpoints,
+    prune_checkpoints,
+)
 
+from hermes_cli.web_routers import hooks as _hooks_routes  # noqa: E402
 
-@app.post("/api/ops/doctor")
-async def run_doctor():
-    try:
-        proc = _spawn_hermes_action(["doctor"], "doctor")
-    except Exception as exc:
-        _log.exception("Failed to spawn doctor")
-        raise HTTPException(status_code=500, detail=f"Failed to run doctor: {exc}")
-    return {"ok": True, "pid": proc.pid, "name": "doctor"}
-
-
-@app.post("/api/ops/security-audit")
-async def run_security_audit():
-    try:
-        proc = _spawn_hermes_action(["security", "audit"], "security-audit")
-    except Exception as exc:
-        _log.exception("Failed to spawn security audit")
-        raise HTTPException(status_code=500, detail=f"Failed to run security audit: {exc}")
-    return {"ok": True, "pid": proc.pid, "name": "security-audit"}
-
-
-def _dashboard_backup_dir() -> Path:
-    return get_hermes_home() / "backups"
-
-
-def _new_dashboard_backup_path() -> Path:
-    stamp = datetime.now().strftime("%Y-%m-%d-%H%M%S")
-    return _dashboard_backup_dir() / f"hermes-backup-{stamp}-{secrets.token_hex(4)}.zip"
-
-
-@app.post("/api/ops/backup")
-async def run_backup(body: BackupRequest):
-    args = ["backup"]
-    archive: Optional[Path] = None
-    output = (body.output or "").strip()
-    if output:
-        args.extend(["-o", output])
-    else:
-        archive = _new_dashboard_backup_path()
-        try:
-            archive.parent.mkdir(parents=True, exist_ok=True)
-        except OSError as exc:
-            raise HTTPException(
-                status_code=500,
-                detail=f"Could not create backup directory: {exc}",
-            )
-        args.extend(["-o", str(archive)])
-    try:
-        proc = _spawn_hermes_action(args, "backup")
-    except Exception as exc:
-        _log.exception("Failed to spawn backup")
-        raise HTTPException(status_code=500, detail=f"Failed to run backup: {exc}")
-    response = {"ok": True, "pid": proc.pid, "name": "backup"}
-    if archive is not None:
-        response["archive"] = str(archive)
-    return response
-
-
-@app.get("/api/ops/backup/download")
-async def download_dashboard_backup(archive: str):
-    try:
-        backup_dir = _dashboard_backup_dir().expanduser().resolve(strict=False)
-        target = Path(archive).expanduser().resolve(strict=True)
-    except FileNotFoundError:
-        raise HTTPException(status_code=404, detail="Backup not found")
-    except (OSError, RuntimeError):
-        raise HTTPException(status_code=400, detail="Invalid backup path")
-
-    if not _path_is_under(backup_dir, target):
-        raise HTTPException(status_code=403, detail="Backup is outside the dashboard backup directory")
-    if not target.is_file():
-        raise HTTPException(status_code=404, detail="Backup not found")
-
-    return FileResponse(
-        path=str(target),
-        media_type="application/zip",
-        filename=target.name,
-        content_disposition_type="attachment",
-    )
-
-
-@app.post("/api/ops/import")
-async def run_import(body: ImportRequest):
-    archive = (body.archive or "").strip()
-    if not archive:
-        raise HTTPException(status_code=400, detail="archive path is required")
-    if not os.path.isfile(archive):
-        raise HTTPException(status_code=404, detail=f"Archive not found: {archive}")
-    args = ["import", archive]
-    if body.force:
-        args.append("--force")
-    try:
-        proc = _spawn_hermes_action(args, "import")
-    except Exception as exc:
-        _log.exception("Failed to spawn import")
-        raise HTTPException(status_code=500, detail=f"Failed to run import: {exc}")
-    return {"ok": True, "pid": proc.pid, "name": "import"}
-
-
-def _safe_backup_upload_name(filename: str | None) -> str:
-    name = Path(filename or "backup.zip").name.strip()
-    name = re.sub(r"[^A-Za-z0-9._-]+", "-", name).strip(".-")
-    if not name:
-        name = "backup.zip"
-    if not name.lower().endswith(".zip"):
-        name = f"{name}.zip"
-    return name
-
-
-@app.post("/api/ops/import-upload")
-async def run_import_upload(
-    file: UploadFile = File(...),
-    force: bool = Form(False),
-):
-    staging_dir = _dashboard_backup_dir()
-    try:
-        staging_dir.mkdir(parents=True, exist_ok=True)
-    except OSError as exc:
-        raise HTTPException(
-            status_code=500,
-            detail=f"Could not create import staging directory: {exc}",
-        )
-
-    safe_name = _safe_backup_upload_name(file.filename)
-    stamp = datetime.now(timezone.utc).strftime("%Y%m%d-%H%M%S")
-    target = staging_dir / f"dashboard-import-{stamp}-{secrets.token_hex(4)}-{safe_name}"
-    tmp_fd, tmp_name = tempfile.mkstemp(
-        prefix=f".{target.name}.",
-        suffix=".upload",
-        dir=str(staging_dir),
-    )
-    tmp_path = Path(tmp_name)
-    total = 0
-    renamed = False
-    try:
-        with os.fdopen(tmp_fd, "wb") as out:
-            while True:
-                chunk = await file.read(_UPLOAD_CHUNK_BYTES)
-                if not chunk:
-                    break
-                total += len(chunk)
-                if total > _MANAGED_FILE_MAX_BYTES:
-                    raise HTTPException(status_code=413, detail="Archive is too large")
-                out.write(chunk)
-        os.replace(tmp_path, target)
-        renamed = True
-    except HTTPException:
-        raise
-    except PermissionError:
-        raise HTTPException(
-            status_code=403,
-            detail="Import staging directory is not writable",
-        )
-    except OSError as exc:
-        raise HTTPException(
-            status_code=500,
-            detail=f"Could not write uploaded archive: {exc}",
-        )
-    finally:
-        if not renamed:
-            tmp_path.unlink(missing_ok=True)
-        await file.close()
-
-    if not zipfile.is_zipfile(target):
-        target.unlink(missing_ok=True)
-        raise HTTPException(
-            status_code=400,
-            detail="Uploaded archive is not a valid zip file",
-        )
-
-    args = ["import", str(target)]
-    if force:
-        args.append("--force")
-    try:
-        proc = _spawn_hermes_action(args, "import")
-    except Exception as exc:
-        _log.exception("Failed to spawn import")
-        raise HTTPException(status_code=500, detail=f"Failed to run import: {exc}")
-    return {
-        "ok": True,
-        "pid": proc.pid,
-        "name": "import",
-        "archive": str(target),
-        "uploaded_bytes": total,
-    }
-
-
-@app.get("/api/ops/hooks")
-async def list_hooks():
-    """List configured shell hooks from config.yaml with consent + health.
-
-    Reports each hook's allowlist (consent) status and whether the script is
-    currently executable, plus the set of valid hook events so the create
-    form can offer them.
-    """
-    from hermes_cli.config import load_config as _load_config
-    from agent import shell_hooks
-
-    try:
-        from hermes_cli.plugins import VALID_HOOKS
-        valid_events = sorted(VALID_HOOKS)
-    except Exception:
-        valid_events = []
-
-    specs = []
-    try:
-        specs = shell_hooks.iter_configured_hooks(_load_config())
-    except Exception:
-        _log.exception("iter_configured_hooks failed")
-
-    out = []
-    for spec in specs:
-        entry = None
-        try:
-            entry = shell_hooks.allowlist_entry_for(spec.event, spec.command)
-        except Exception:
-            pass
-        executable = False
-        try:
-            executable = shell_hooks.script_is_executable(spec.command)
-        except Exception:
-            pass
-        out.append({
-            "event": spec.event,
-            "matcher": spec.matcher,
-            "command": spec.command,
-            "timeout": spec.timeout,
-            "allowed": entry is not None,
-            "approved_at": (entry or {}).get("approved_at"),
-            "executable": executable,
-        })
-
-    return {"hooks": out, "valid_events": valid_events}
-
-
-@app.post("/api/ops/hooks")
-async def create_hook(body: HookCreate):
-    """Add a shell hook to config.yaml (and optionally approve it).
-
-    Shell hooks run arbitrary commands, so this is a privileged action: it
-    writes to the ``hooks:`` config block and, when ``approve`` is set, records
-    consent in the allowlist so the hook actually fires.  Takes effect on the
-    next session / gateway restart.
-    """
-    from agent import shell_hooks
-
-    event = (body.event or "").strip()
-    command = (body.command or "").strip()
-    if not event or not command:
-        raise HTTPException(status_code=400, detail="event and command are required")
-
-    try:
-        from hermes_cli.plugins import VALID_HOOKS
-        if event not in VALID_HOOKS:
-            raise HTTPException(
-                status_code=400,
-                detail=f"Unknown event '{event}'. Valid: {', '.join(sorted(VALID_HOOKS))}",
-            )
-    except HTTPException:
-        raise
-    except Exception:
-        pass
-
-    cfg = load_config()
-    hooks_cfg = cfg.get("hooks")
-    if not isinstance(hooks_cfg, dict):
-        hooks_cfg = {}
-        cfg["hooks"] = hooks_cfg
-    entries = hooks_cfg.get(event)
-    if not isinstance(entries, list):
-        entries = []
-        hooks_cfg[event] = entries
-
-    new_entry: Dict[str, Any] = {"command": command}
-    if body.matcher:
-        new_entry["matcher"] = body.matcher
-    if body.timeout is not None:
-        new_entry["timeout"] = int(body.timeout)
-    entries.append(new_entry)
-    save_config(cfg)
-
-    approved = False
-    if body.approve:
-        try:
-            shell_hooks._record_approval(event, command)
-            approved = True
-        except Exception:
-            _log.exception("hook consent record failed")
-
-    return {"ok": True, "event": event, "command": command, "approved": approved}
-
-
-@app.delete("/api/ops/hooks")
-async def delete_hook(body: HookDelete):
-    """Remove a hook from config.yaml and revoke its consent allowlist entry."""
-    from agent import shell_hooks
-
-    event = (body.event or "").strip()
-    command = (body.command or "").strip()
-    if not event or not command:
-        raise HTTPException(status_code=400, detail="event and command are required")
-
-    cfg = load_config()
-    hooks_cfg = cfg.get("hooks")
-    removed = False
-    if isinstance(hooks_cfg, dict) and isinstance(hooks_cfg.get(event), list):
-        before = len(hooks_cfg[event])
-        hooks_cfg[event] = [
-            e for e in hooks_cfg[event]
-            if not (isinstance(e, dict) and e.get("command") == command)
-        ]
-        removed = len(hooks_cfg[event]) < before
-        if not hooks_cfg[event]:
-            del hooks_cfg[event]
-        if not hooks_cfg:
-            cfg.pop("hooks", None)
-        save_config(cfg)
-
-    # Revoke consent regardless so a re-add re-prompts.
-    try:
-        shell_hooks.revoke(command)
-    except Exception:
-        pass
-
-    if not removed:
-        raise HTTPException(status_code=404, detail="No matching hook found")
-    return {"ok": True}
-
-
-@app.get("/api/ops/checkpoints")
-async def list_checkpoints():
-    """List the /rollback shadow store checkpoints (read-only)."""
-    # Checkpoints live under <hermes_home>/checkpoints/.  Surface a count +
-    # total size so the dashboard can show what a prune would reclaim; the
-    # actual prune is a spawned action so confirmation/pruning logic stays
-    # in one place (the CLI).
-    cp_dir = get_hermes_home() / "checkpoints"
-    sessions = []
-    total_bytes = 0
-    if cp_dir.is_dir():
-        for child in sorted(cp_dir.iterdir()):
-            if not child.is_dir():
-                continue
-            size = 0
-            count = 0
-            for f in child.rglob("*"):
-                if f.is_file():
-                    try:
-                        size += f.stat().st_size
-                        count += 1
-                    except OSError:
-                        pass
-            total_bytes += size
-            sessions.append({
-                "session": child.name,
-                "files": count,
-                "bytes": size,
-            })
-    return {"sessions": sessions, "total_bytes": total_bytes}
-
-
-@app.post("/api/ops/checkpoints/prune")
-async def prune_checkpoints():
-    try:
-        proc = _spawn_hermes_action(["checkpoints", "prune"], "checkpoints-prune")
-    except Exception as exc:
-        _log.exception("Failed to spawn checkpoints prune")
-        raise HTTPException(status_code=500, detail=f"Failed to prune checkpoints: {exc}")
-    return {"ok": True, "pid": proc.pid, "name": "checkpoints-prune"}
+app.include_router(_hooks_routes.router)
+from hermes_cli.web_routers.hooks import (  # noqa: E402,F401 — legacy re-exports; tests call these via web_server.<name>
+    list_hooks,
+    create_hook,
+    delete_hook,
+)
 
 
 # ---------------------------------------------------------------------------
